Implement basic tracing game mechanism
NOTES:
The algorithm to calculate percentage is insanely slow and blows up the player slightly
We should probably only be checking new pixels, which still won't be great, but better.
Ideally, we would never add redundant pixels to the graph
perhaps we could turn this into a lookup... Maybe make the pixel graph a giant hash?
We'll do it if we have to